Tampon, rejet sélectif

+++ Hamming.h
@@ -2,8 +2,8 @@
 #include <stdint.h> 
 namespace Hamming
 {
-	bool detecter(uint16_t donnees);
-	uint16_t corriger(uint16_t donnees);
-	uint16_t encoder(uint16_t donnees);
-	uint16_t decoder(uint16_t donnees);
+	bool detecter(uint32_t donnees);
+	uint32_t corriger(uint32_t donnees);
+	uint32_t encoder(uint32_t donnees);
+	uint32_t decoder(uint32_t donnees);
 }
